+++ server/app/auth.py
@@ -1,15 +1,18 @@
 import functools
 from datetime import datetime, timedelta
+import asyncio
+
 from jose import jwt, JWTError
 from typing import Optional
-from .response_exceptions import CredentialsException
-from .config.token_conf import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES
-from .redis_req import add_revoked_redis, is_member_revoked_redis
 from pydantic import BaseModel
 from fastapi import Depends
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 import boto3
-import asyncio
+
+from .response_exceptions import CredentialsException
+from .config.token_conf import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES
+from .redis_req import add_revoked_redis, is_member_revoked_redis
+
 
 security = HTTPBearer(scheme_name='bearer')
 
@@ -61,12 +64,12 @@ def request_temp_cred(session_obj: boto3.Session, region: str) -> dict:
     return cred
 
 
-async def request_temp_cred_async(session_obj: boto3.Session, region: str) -> dict:
+async def async_request_temp_cred(session_obj: boto3.Session, region: str) -> dict:
     fun = functools.partial(request_temp_cred, session_obj=session_obj, region=region)
     return await asyncio.to_thread(fun)
 
 
-def temp_session_create(cred: dict) -> boto3.Session:
+def create_temp_session(cred: dict) -> boto3.Session:
     return boto3.Session(aws_access_key_id=cred['AccessKeyId'],
                          aws_secret_access_key=cred['SecretAccessKey'],
                          aws_session_token=cred['SessionToken'],
@@ -81,15 +84,4 @@ async def validate_and_decode_access_token(auth_header: HTTPAuthorizationCredent
     except JWTError:
         raise CredentialsException
     return payload
-
-
-async def get_user_id(decoded: dict = Depends(validate_and_decode_access_token)) -> str:
-    """
-    토큰에서 유저 id를 가져옴. 토큰이 유효하지 않을 경우 에러 raise
-    """
-    try:
-        user_id: str = decoded['user_id']
-    except KeyError:
-        raise CredentialsException
-    return user_id
